+++ fsl_mrs/models/model_freeshift.py
@@ -11,14 +11,14 @@ from fsl_mrs.utils.misc import FIDToSpec
 from fsl_mrs.models.model_voigt import _init_params_voigt
 
 
-def vars(n_basis, n_groups, b_order):
+def vars(n_basis, n_groups, n_baseline):
     """Return the names and sizes of each parameter in the model
     :param n_basis: Number of basis spectra
     :type n_basis: int
     :param n_groups: Number of metabolite groups
     :type n_groups: int
-    :param b_order: Baseline polynomial order
-    :type b_order: int
+    :param n_baseline: Number baseline bases
+    :type n_baseline: int
     :return: List of parameter names
     :rtype: List
     :return: List of parameter sizes
@@ -32,19 +32,19 @@ def vars(n_basis, n_groups, b_order):
         n_basis,  # eps
         1,  # Phi_0
         1,  # Phi_1
-        2 + (b_order * 2)]  # baseline
+        n_baseline * 2]  # baseline
     return var_names, sizes
 
 
-def bounds(n_basis, n_groups, b_order, method, disableBaseline=False):
+def bounds(n_basis, n_groups, n_baseline, method, disableBaseline=False):
     """Return bounds for the methods
 
     :param n_basis: Number of basis spectra
     :type n_basis: int
     :param n_groups: Number of metabolite groups
     :type n_groups: int
-    :param b_order: Polynomial baseline order
-    :type b_order: int
+    :param n_baseline: Number baseline bases
+    :type n_baseline: int
     :param method: Fitting optimisation method. 'Newton' or 'MH'
     :type method: str
     :param disableBaseline: Disable the baseline by seting bounds to 0, defaults to False
@@ -65,7 +65,7 @@ def bounds(n_basis, n_groups, b_order, method, disableBaseline=False):
         # phi0, phi1
         bnds.extend([(None, None)] * 2)
         # baseline
-        n = (1 + b_order) * 2
+        n = n_baseline * 2
         if disableBaseline:
             bnds.extend([(0.0, 0.0)] * n)
         else:
@@ -91,7 +91,7 @@ def bounds(n_basis, n_groups, b_order, method, disableBaseline=False):
         LB.extend([MIN] * 2)
         UB.extend([MAX] * 2)
         # baseline
-        n = (1 + b_order) * 2
+        n = n_baseline * 2
         if disableBaseline:
             LB.extend([0.0] * n)
             UB.extend([0.0] * n)
@@ -105,7 +105,7 @@ def bounds(n_basis, n_groups, b_order, method, disableBaseline=False):
 def mask(
         n_basis,
         n_groups,
-        b_order,
+        n_baseline,
         fit_conc=True,
         fit_shape=True,
         fit_phase=True,
@@ -116,8 +116,8 @@ def mask(
     :type n_basis: int
     :param n_groups: Number of metabolite groups
     :type n_groups: int
-    :param b_order: Polynomial baseline order
-    :type b_order: int
+    :param n_baseline: Number baseline bases
+    :type n_baseline: int
     :param fit_conc: Whether to fit the concentrations, defaults to True
     :type fit_conc: bool, optional
     :param fit_shape: Whether to fit the lineshapes, defaults to True
@@ -143,7 +143,7 @@ def mask(
         mask.extend([1] * 2)
     else:
         mask.extend([0] * 2)
-    n = (1 + b_order) * 2
+    n = n_baseline * 2
     if fit_baseline:
         mask.extend([1] * n)
     else: