+++ fsl_mrs/scripts/fsl_mrs_summarise.py
@@ -22,7 +22,7 @@ from fsl_mrs.utils import results
 from fsl_mrs.utils import mrs_io
 from fsl_mrs.utils import misc
 from fsl_mrs.utils import plotting
-from fsl_mrs.utils.baseline import prepare_baseline_regressor
+from fsl_mrs.utils.baseline import Baseline
 
 
 def main():
@@ -150,21 +150,31 @@ def main():
         # Generate metabolite groups
         metab_groups = misc.parse_metab_groups(mrs, orig_args['metab_groups'])
         baseline_order = orig_args['baseline_order']
-        if baseline_order < 0:
-            baseline_order = 0  # Required to make prepare_baseline_regressor run.
+        if 'baseline' in orig_args:
+            baseline_arg = orig_args['baseline']
+        else:
+            baseline_arg = None
         ppmlim = orig_args['ppmlim']
         if ppmlim is None:
             from fsl_mrs.utils.constants import nucleus_constants
             ppmlim = nucleus_constants(mrs.nucleus).ppm_range
 
-        # Generate baseline polynomials (B)
-        B = prepare_baseline_regressor(mrs, baseline_order, ppmlim)
+        # Generate baseline object
+        baseline_obj = Baseline(
+            mrs,
+            ppmlim,
+            baseline_arg,
+            baseline_order)
 
         # Generate results object
         res_store[name] = results.FitRes(
             mrs,
             param_df['mean'].to_numpy(),
-            model, method, metab_groups, baseline_order, B, ppmlim,
+            model,
+            method,
+            metab_groups,
+            baseline_obj,
+            ppmlim,
             runqc=False)
 
         if orig_args['combine'] is not None: