+++ fsl_mrs/tests/test_models.py
@@ -45,6 +45,7 @@ def test_getInit():
 n_basis = 10
 n_groups = 2
 b_order = 2
+n_baseline = b_order + 1
 
 answer_names = []
 answer_sizes = []
@@ -64,7 +65,7 @@ answer_sizes.append([10, 2, 2, 10, 1, 1, 6])
 
 def test_FSLModel_vars():
     for model, ans_n, ans_s in zip(all_models, answer_names, answer_sizes):
-        names, sizes = models.FSLModel_vars(model, n_basis, n_groups, b_order)
+        names, sizes = models.FSLModel_vars(model, n_basis, n_groups, n_baseline)
         assert names == ans_n
         assert sizes == ans_s
 
@@ -72,18 +73,18 @@ def test_FSLModel_vars():
 def test_FSLModel_bounds():
     for model, ans_s in zip(all_models, answer_sizes):
         # Newton
-        bounds = models.FSLModel_bounds(model, n_basis, n_groups, b_order, 'Newton')
+        bounds = models.FSLModel_bounds(model, n_basis, n_groups, n_baseline, 'Newton')
         assert len(bounds) == sum(ans_s)
 
         # MH
-        LB, UB = models.FSLModel_bounds(model, n_basis, n_groups, b_order, 'MH')
+        LB, UB = models.FSLModel_bounds(model, n_basis, n_groups, n_baseline, 'MH')
         assert len(LB) == sum(ans_s)
         assert len(UB) == sum(ans_s)
 
 
 def test_FSLModel_mask():
     for model, ans_s in zip(all_models, answer_sizes):
-        mask = models.FSLModel_mask(model, n_basis, n_groups, b_order)
+        mask = models.FSLModel_mask(model, n_basis, n_groups, n_baseline)
         # Final 6 (baseline) parameters should be zero
         assert sum(mask[:-6]) == len(mask[:-6])
         assert sum(mask[-6:]) == 0